refactor: Restructure CLI tool into organized, maintainable architecture
Refactored Program.cs from 1,613 lines to 85 lines by extracting code into
well-organized class files following proper naming conventions and SOLID principles.

New structure:
- Configuration/: API configuration and constants
- Models/: Data models (CommandLineOptions, API request/response models)
- Services/: Core business logic (ApiClient, PackageSelector, ScriptExecutor)
- UI/: Console display and user interaction (Display, Prompts, Tables)
- Workflows/: Mode orchestration (InteractiveModeWorkflow, CliModeWorkflow)
- Program.cs: Minimal entry point for initialization and delegation

Benefits:
- Single Responsibility Principle: Each class has one clear purpose
- Separation of Concerns: UI, logic, and API communication separated
- Improved maintainability and testability
- Better code organization and discoverability
- No behavioral changes - pure refactoring
