+++ src/PackageCliTool/Models/Api/PackageVersionResponse.cs
@@ -1,0 +1,12 @@
+using System.Text.Json.Serialization;
+
+namespace PackageCliTool.Models.Api;
+
+/// <summary>
+/// Response model for package versions from the API
+/// </summary>
+public class PackageVersionResponse
+{
+    [JsonPropertyName("versions")]
+    public List<string> Versions { get; set; } = new();
+}
